refactor(codegen): extract locus dissolve helpers into locus/dissolve.rs (Round 4a)
8 fns lifted into LocusDissolve<'ctx>: lower_locus_kmax,
lower_locus_draining, emit_method_return_deep_copy,
emit_locus_field_dissolves, emit_locus_field_owned_branch,
emit_locus_field_drains, emit_birth_check, emit_locus_arena_destroy.

These run at dissolve / drain time across the locus's fields +
emit the m43 / k_max / draining synthetic-field reads. Round 4
focuses on locus codegen and this is the smallest self-contained
piece to lift first.

Surface bumps to pub(crate):
- Cx fields: user_loci, main_placement_map, current_method_caller_arena,
  main_locus_name
- Cx methods: emit_cross_arena_store_deep_copy_ptr,
  emit_return_value_deep_copy, resolve_failure_route,
  ty_needs_self_field_deep_copy
- LocusInfo struct + all fields (bulk-bumped — Round 4 sub-rounds
  ahead will need them too)
- CapacitySlotLayout struct + fields
- SlotForm enum

Imports CapacitySlotKind / ProjectionClass / RecognitionSubMode /
ScheduleClass / BirthCheckDecl from hale_syntax::ast.

codegen.rs: 35071 -> 33895 (-1176 lines).
