+++ crates/hale-codegen/src/locus/mod.rs
@@ -1,0 +1,4 @@
+//! Locus codegen (lifecycle, methods, dissolve, arenas, placement,
+//! return path). Round 4 of the codegen model-org refactor.
+
+pub(crate) mod dissolve;
